+++ app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:attrs="http://schemas.android.com/apk/res-auto"
-    tools:context=".MainActivity"
+    tools:context=".Actividades.MainActivity"
     android:id="@+id/drawerll"
     android:fitsSystemWindows="true">
