+++ app/src/main/res/menu/menu_search.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools" tools:context=".searchableactivity">
+    xmlns:tools="http://schemas.android.com/tools" tools:context=".Actividades.searchableactivity">
     <item
         android:id="@+id/procura"
         android:title="@string/app_hint"
